Removes more magic numbers
Always a good thing
basically determines what exactly a second is in code rather than just
assuming it's 72 cuz idk if it changes depending on computer quality.

I can guarentee it would break if it's any larger 72 and would be
significantly slower if it was less, and different tiers of comuputer
might affect that. Thus we just find it